+++ src/main/java/webserver/controller/UserCreateController.java
@@ -3,11 +3,11 @@ package webserver.controller;
 import db.DataBase;
 import model.User;
 import webserver.exception.ApiException;
-import webserver.request.HttpMethod;
-import webserver.request.HttpRequest;
-import webserver.request.Path;
-import webserver.response.HttpResponse;
-import webserver.response.HttpStatusCode;
+import webserver.http.request.HttpMethod;
+import webserver.http.request.HttpRequest;
+import webserver.http.request.Path;
+import webserver.http.response.HttpResponse;
+import webserver.http.response.HttpStatusCode;
 
 public class UserCreateController implements Controller {
