+++ src/main/java/webserver/http/HttpHeaders.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.http;
 
 public final class HttpHeaders {
