+++ src/main/java/webserver/http/HttpSession.java
@@ -1,0 +1,4 @@
+package webserver.http;
+
+public class HttpSession {
+}
