+++ src/main/java/webserver/http/request/QueryString.java
@@ -1,4 +1,4 @@
-package webserver.request;
+package webserver.http.request;
 
 import utils.UrlQueryUtils;
