+++ src/main/java/webserver/http/request/RequestLine.java
@@ -1,4 +1,4 @@
-package webserver.request;
+package webserver.http.request;
 
 import utils.Assert;
