+++ src/main/java/webserver/http/response/ResponseHeader.java
@@ -1,4 +1,4 @@
-package webserver.response;
+package webserver.http.response;
 
 import utils.Assert;
