+++ src/test/java/webserver/http/request/RequestLineTest.java
@@ -1,4 +1,4 @@
-package webserver.request;
+package webserver.http.request;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
