+++ src/test/java/webserver/http/response/HttpResponseTest.java
@@ -1,4 +1,4 @@
-package webserver.response;
+package webserver.http.response;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
