+++ src/test/java/webserver/step/UserAcceptanceStep.java
@@ -5,7 +5,7 @@ import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import webserver.HttpHeaders;
+import webserver.http.HttpHeaders;
 
 import java.util.Map;
